feat: INFRA-160 Moved sync filesystem function in separate file
